[CHORE] 사용할 수 있는 닉네임인지 확인하는 API 위치 이동 (Auth -> Member)
