+++ src/main/java/com/service/ttucktak/dto/member/GetNicknameAvailableResDto.java
@@ -1,4 +1,4 @@
-package com.service.ttucktak.dto.auth;
+package com.service.ttucktak.dto.member;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
